feat: 입력받는 User클래스 생성, tryCnt 입력 및 검증
